+++ internal/domain/network.go
@@ -82,12 +82,14 @@ type UpdateNetworkObject struct {
 }
 
 type NetworkObjectFilters struct {
-	AccountID  int64
-	Provider   string
-	Region     string
-	ObjectType string
-	State      string
-	Query      string
+	AccountID          int64
+	Provider           string
+	Region             string
+	ObjectType         string
+	State              string
+	PoolID             int64
+	SourceDiscoveredID string
+	Query              string
 }
 
 type NetworkObjectListResponse struct {